Adds UI for future predictions
