+++ components/Menu.tsx
@@ -64,7 +64,10 @@ const Layout = ({ league }: MainInterface) => {
   // Checks if the player should see the league links
   if (league) {
     pages.push({ name: t("Standings"), link: `/${league}` });
-    pages.push({ name: t("Predictions"), link: `/${league}/predictions` });
+    pages.push({
+      name: t("Predictions"),
+      link: `/${league}/${session?.user.id}/predictions`,
+    });
     pages.push({ name: t("Squad"), link: `/${league}/squad` });
     pages.push({ name: t("Transfers"), link: `/${league}/transfer` });
   }
